Outsourced Types class
