+++ pokete_classes/classes.py
@@ -1,7 +1,6 @@
 import os
 import scrap_engine as se
 from pokete_classes.ui_elements import StdFrame
-from pokete_classes.color import Color
 
 
 class PlayMap(se.Map):
@@ -25,24 +24,6 @@ class PlayMap(se.Map):
         """Executes the extra action"""
         if self.__extra_actions is not None:
             self.__extra_actions()
-
-
-class Types:
-    """Class to organize PokeTypes"""
-
-    def __init__(self, types):
-        for i in types:
-            setattr(self, i, PokeType(i, **types[i]))
-
-
-class PokeType():
-    """Type for Poketes and attacks"""
-
-    def __init__(self, name, effective, ineffective, color):
-        self.name = name
-        self.effective = effective
-        self.ineffective = ineffective
-        self.color = "" if color is None else str.join("", [getattr(Color, i) for i in color])
 
 
 class Settings():
